// frontend/src/sdk/generated/types.gen.ts
export type Game = {
  id: string;
  name: string;
  title: string;
  developer: string;
  description?: string;
  category: string;
  tags: Array<string>;
  thumbnailUrl?: string;
  bannerUrl?: string;
  isActive?: boolean;
};
// ...
export type PostApiGamesByIdEnterData = {
  body?: never;
  path: {
    id: string;
  };
  query?: never;
  url: "/api/games/{id}/enter";
};

export type PostApiGamesByIdEnterErrors = {
  /**
   * Not Found
   */
  404: {
    message: string;
  };
};

export type PostApiGamesByIdEnterError =
  PostApiGamesByIdEnterErrors[keyof PostApiGamesByIdEnterErrors];

export type PostApiGamesByIdEnterResponses = {
  /**
   * Game session details
   */
  200: {
    webUrl: string;
    gameConfig: {
      authToken: string;
      gameSessionId: string;
      userId: string;
      gameName: string;
      lang: string;
      currency: string;
      operator: string;
      provider: string;
      depositUrl: string;
      lobbyUrl: string;
      mode: string;
      rgsApiBase: string;
      cdn: string;
      baseCdn: string;
    };
  };
};

export type PostApiGamesByIdEnterResponse =
  PostApiGamesByIdEnterResponses[keyof PostApiGamesByIdEnterResponses];